feat: Refactored MediaPipe repo to have possibility to use face mesh as static library
