+++ .bazelrc
@@ -57,7 +57,7 @@ build:android_arm64 --cpu=arm64-v8a
 build:android_arm64 --fat_apk_cpu=arm64-v8a
 
 # iOS configs.
-build:ios --apple_platform_type=ios
+build:ios --apple_platform_type=ios --define ios_config=ios_enable_gpu
 
 build:ios_i386 --config=ios
 build:ios_i386 --cpu=ios_i386
@@ -74,6 +74,10 @@ build:ios_armv7 --watchos_cpus=armv7k
 build:ios_arm64 --config=ios
 build:ios_arm64 --cpu=ios_arm64
 build:ios_arm64 --watchos_cpus=armv7k
+
+build:ios_disable_gpu_arm64 --config=ios_arm64
+build:ios_disable_gpu_arm64 --define ios_config=ios_disable_gpu
+build:ios_disable_gpu_arm64 --define MEDIAPIPE_DISABLE_GPU=1
 
 build:ios_arm64e --config=ios
 build:ios_arm64e --cpu=ios_arm64e
