+++ mediapipe/calculators/core/gate_calculator.h
