+++ mediapipe/calculators/image/image_properties_calculator.h
