+++ mediapipe/calculators/tensor/image_to_tensor_calculator.h
