+++ mediapipe/calculators/util/local_file_contents_calculator.h
